Move toast files into toast namespace
