+++ Dalamud/Game/Gui/GameGui.cs
@@ -5,6 +5,7 @@ using System.Runtime.InteropServices;
 using Dalamud.Game.Gui.Addons;
 using Dalamud.Game.Gui.FlyText;
 using Dalamud.Game.Gui.PartyFinder;
+using Dalamud.Game.Gui.Toast;
 using Dalamud.Game.Text.SeStringHandling.Payloads;
 using Dalamud.Hooking;
 using Dalamud.Interface;
