+++ Dalamud/Game/Gui/Toast/ToastGui.cs
@@ -2,11 +2,10 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-using Dalamud.Game.Gui.Toast;
 using Dalamud.Game.Text.SeStringHandling;
 using Dalamud.Hooking;
 
-namespace Dalamud.Game.Gui
+namespace Dalamud.Game.Gui.Toast
 {
     /// <summary>
     /// This class facilitates interacting with and creating native toast windows.
